django_auth/requirements.txt
asgiref==3.7.2
Django==4.2
django-autoslug==1.9.9
django-phonenumber-field==7.3.0
phonenumbers==8.13.30
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.9.0
